backend/requirements.txt

fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.5.2
redis==5.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pybase62==1.0.0
alembic==1.13.3
bcrypt==4.0.1